requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
